a function to calculate thermal time

this function calculates the thermal time for an experiment according to several methods

thermalTime(
  datain,
  inSpecie,
  method,
  inDateS = NULL,
  inDateE = NULL,
  inTemp = NULL
)

Arguments

datain
input dataframe of meteo data from phis web service

inSpecie
character, studied specie

method
character, a method of thermal time's calculation ("parent","baseline")

inDateS
a date of sowing or thinning etc... start event ("YYYY-MM-DD")

inDateE
a date of harvesting ("YYYY-MM-DD")

inTemp
numeric, a baseline temperature for baseline's method

Value

a dataframe

Details

Parent et. al. (2010). Modeling temperature-compensated physiological rates, based on the co-ordination of responses to temperature of developmental processes. Journal of Experimental Botany. 61 (8):2057-2069

if the Parent's model is chosen inTemp must be NULL

The input dataframe is extracted from phis web service (getEnvironment function) and is structured as follow: date, value, sensor, codeVariable and facility

Examples

# Example for the model of Parent 2010
data(meteoDT)
head(meteoDT)

#>                  date      yyyymmdd
#> 1 2017-03-29 22:00:00 mars 29, 2017
#> 2 2017-03-29 22:15:00 mars 29, 2017
#> 3 2017-03-29 22:30:00 mars 29, 2017
#> 4 2017-03-29 22:45:00 mars 29, 2017
#> 5 2017-03-29 23:00:00 mars 29, 2017
#> 6 2017-03-29 23:15:00 mars 29, 2017
#>                                          codeVariable value
#> 1 leaf temperature_thermocouple sensor_degree celsius    21
#> 2 leaf temperature_thermocouple sensor_degree celsius    21
#> 3 leaf temperature_thermocouple sensor_degree celsius    21
#> 4 leaf temperature_thermocouple sensor_degree celsius    21
#> 5 leaf temperature_thermocouple sensor_degree celsius    20
#> 6 leaf temperature_thermocouple sensor_degree celsius    20

test<-thermalTime(datain=meteoDT,inSpecie="maize",method="parent",
                  inDateS="2017-04-02",
                  inDateE="2017-06-15",inTemp=NULL)
head(test)

#>          Day       TT
#> 1 2017-04-02 1.006146
#> 2 2017-04-03 2.139710
#> 3 2017-04-04 3.236158
#> 4 2017-04-05 4.214921
#> 5 2017-04-06 5.269632
#> 6 2017-04-07 6.426939